+++ transport-udt/src/main/java/io/netty/channel/udt/nio/NioUdtAcceptorChannel.java
@@ -17,10 +17,13 @@ package io.netty.channel.udt.nio;
 
 import com.barchart.udt.TypeUDT;
 import com.barchart.udt.nio.ServerSocketChannelUDT;
+import com.barchart.udt.nio.SocketChannelUDT;
 import io.netty.channel.ChannelException;
+import io.netty.channel.ChannelMetadata;
 import io.netty.channel.ChannelOutboundBuffer;
 import io.netty.channel.nio.AbstractNioMessageChannel;
 import io.netty.channel.udt.DefaultUdtServerChannelConfig;
+import io.netty.channel.udt.UdtChannel;
 import io.netty.channel.udt.UdtServerChannel;
 import io.netty.channel.udt.UdtServerChannelConfig;
 import io.netty.util.internal.logging.InternalLogger;
@@ -28,6 +31,7 @@ import io.netty.util.internal.logging.InternalLoggerFactory;
 
 import java.net.InetSocketAddress;
 import java.net.SocketAddress;
+import java.util.List;
 
 import static java.nio.channels.SelectionKey.*;
 
@@ -38,6 +42,8 @@ public abstract class NioUdtAcceptorChannel extends AbstractNioMessageChannel im
 
     protected static final InternalLogger logger =
             InternalLoggerFactory.getInstance(NioUdtAcceptorChannel.class);
+
+    private static final ChannelMetadata METADATA = new ChannelMetadata(false);
 
     private final UdtServerChannelConfig config;
 
@@ -132,4 +138,21 @@ public abstract class NioUdtAcceptorChannel extends AbstractNioMessageChannel im
         return null;
     }
 
+    @Override
+    public ChannelMetadata metadata() {
+        return METADATA;
+    }
+
+    @Override
+    protected int doReadMessages(List<Object> buf) throws Exception {
+        final SocketChannelUDT channelUDT = javaChannel().accept();
+        if (channelUDT == null) {
+            return 0;
+        } else {
+            buf.add(newConnectorChannel(channelUDT));
+            return 1;
+        }
+    }
+
+    protected abstract UdtChannel newConnectorChannel(SocketChannelUDT channelUDT);
 }
